added cat of ios build + android sim command
I'm getting an error that android isn't being simulated even though I am 90% sure Ionic handles the simulation itself.  I'm going to run a build with it in and see what happens.  Also I want to see what's in the log for the ios sim and make sure it isn't erroring silently