change shebang in before_scripts [ci skip]
changed to bash.  Other scripts were in bash and there is a less likely chance that changing sh to bash will cause commands to no longer work.  Also the efficiency loss on the micro scale doesn't matter when builds are taking many minutes and are executed not that often.  I believe the robustness and flexibility of bash would provide a better performance increase.